refactor: customer 패키지 생성
